pyproject.toml
[tool.poetry]
name = "muzloc"
version = "0.1.0"
description = ""
authors = ["Aleksandr <aleksandr.voitushenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
psycopg2-binary = "^2.9.9"
faker = "^25.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
